+++ src/apps/novadesk/shared/Settings.h
@@ -107,7 +107,7 @@ private:
   static json s_Data;          ///< In-memory settings data.
   static bool s_Dirty;         ///< True if unsaved changes exist.
   static bool s_IsFirstRun;    ///< True if no settings file was found on init.
-  static DWORD s_LastSaveTick; ///< Tick count of last save (for debounce).
+  static ULONGLONG s_LastSaveTick; ///< Tick count of last save (for debounce).
 };
 
 #endif
